+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: rnovotny <rnovotny@student.42prague.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/10 12:50:28 by rnovotny          #+#    #+#             */
-/*   Updated: 2023/01/16 13:42:17 by rnovotny         ###   ########.fr       */
+/*   Updated: 2023/01/19 16:23:36 by rnovotny         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
 	char	*substr;
-	int		i;
+	size_t	i;
 
 	substr = (char *)malloc((len + 1) * sizeof(char));
 	if (substr == 0)
